fix all ps2 warning
git-svn-id: https://twin-e.svn.sourceforge.net/svnroot/twin-e/trunk@145 b6842851-fe11-0410-aa3d-efb4359767e1

+++ twin-e/src/renderer.c
@@ -1269,7 +1269,7 @@ void FillVertic_A(int ecx, int edi)
           {
             unsigned short int ax;
             bh ^= 1;
-            ax = (unsigned short int) out2;
+            ax = (unsigned short int)(int) out2;
             ax &= 1;
             if(ax ^ bh)
             {
